Add step_calc function to ArithmeticExpression trait
This change removed the expression_manipulator.rs file and added a step_calc function defined in ArithmeticExpression. This function, which calculates the next step (and indicate if it does so), is implemented in the Number, Bracket, and BinaryOperation structs which are all part of the 'syntax' module. Tests were updated to ensure all functions are working correctly with the newly added functionality.