Expose type_descriptor elements from lib.rs
The elements from the type_descriptor module are now publically exposed from the lib.rs. Additionally, substantial changes have been made in the test case for checking the descriptions of point structure. Tests are now aiming to be more precise and robust with specific assertions and error handling in place.